refactor and update rekor dep
